+++ src/game/nlcore/action/actionTypes.ts
@@ -5,13 +5,13 @@ import {CommonDisplayable, ImageColor, ImageSrc} from "@core/types";
 import {Transform} from "@core/elements/transform/transform";
 import type {Scene} from "@core/elements/scene";
 import type {MenuData} from "@core/elements/menu";
-import {Awaitable} from "@lib/util/data";
-import {ITransition} from "@core/elements/transition/type";
+import {Awaitable, FlexibleTuple, SelectElementFromEach} from "@lib/util/data";
+import {IImageTransition, ITransition} from "@core/elements/transition/type";
 import type {Sound} from "@core/elements/sound";
 import type {Script} from "@core/elements/script";
 import {Sentence} from "@core/elements/character/sentence";
 import type {TransformDefinitions} from "@core/elements/transform/type";
-import {Image} from "@core/elements/image";
+import {Image, TagDefinition} from "@core/elements/image";
 
 /* Character */
 export const CharacterActionTypes = {
@@ -81,6 +81,7 @@ export const ImageActionTypes = {
     applyTransition: "image:applyTransition",
     flush: "image:flush",
     initWearable: "image:initWearable",
+    setAppearance: "image:setAppearance",
 } as const;
 export type ImageActionContentType = {
     [K in typeof ImageActionTypes[keyof typeof ImageActionTypes]]:
@@ -95,7 +96,8 @@ export type ImageActionContentType = {
                                     K extends "image:applyTransition" ? [ITransition] :
                                         K extends "image:flush" ? [] :
                                             K extends "image:initWearable" ? [Image] :
-                                                any;
+                                                K extends "image:setAppearance" ? [FlexibleTuple<SelectElementFromEach<TagDefinition>>, IImageTransition | undefined] :
+                                                    any;
 }
 /* Condition */
 export const ConditionActionTypes = {